Trying a global lobby idea
It seems when players log in and click online play a lobby is created and instead I'm attempting to create a global lobby that will be active one a singular player logs in. It doesn't matter if 100 players login a singular lobby will exist on the server. Future implementation would involve lobbies being created as more players are attempting to play online.